:recycle:Refactor: refactor backtest result code split chart 관련 함수 분리 styles & types 관련 코드 분리
