+++ tradin/src/components/units/backtest/main/BackTest.styles.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { FaRocket } from "react-icons/fa";
-import { css} from "@emotion/react";
+import { css } from "@emotion/react";
 
 export const Container = styled.div`
     display: flex;
@@ -204,111 +204,4 @@ export const BackTestButton = styled.button`
 `
 
 export const StyledRocketIcon = styled(FaRocket)`
-`;
-
-export const ResultContainer = styled.div`
-    background-color: ${({ theme }) => theme.innerbackgroundColor};
-    width: 100%;
-    flex-grow: 1;
-    padding: 2rem;
-    border-radius: 8px;
-    overflow: auto;
-    margin-bottom: 1rem;
-`;
-
-export const ResultHeader = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    margin-bottom: 1rem;
-    width: 100%;
-`;
-
-export const ExecutedOptions = styled.span`
-    font-size: 0.9em;
-    color: ${({ theme }) => theme.timeTextColor};
-`;
-
-export const ResultInnerContainer = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-    margin-bottom: 1rem;
-`
-
-export const ResultTitle = styled.h4`
-    color: ${({ theme }) => theme.textColor};
-    margin-right: 1rem;
-    transition: all 0.3s ease;
-    width: auto;
-`;
-
-export const ResultSubtitle = styled.h5`
-    color: ${({ theme }) => theme.textColor};
-    font-size: 1.1em;
-    font-weight: 600;
-    margin-top: 1.5rem;
-    margin-bottom: 1rem;
-    padding-bottom: 0.5rem;
-    border-bottom: 1px solid ${({ theme }) => theme.borderColor};
-    width: 100%;
-    transition: all 0.3s ease;
-`;
-
-export const ResultContent = styled.p`
-    color: ${({ theme }) => theme.textColor};
-    width: calc(50% - 0.5rem);
-    margin: 0;
-    transition: all 0.3s ease;
-    display: flex;
-    align-content: center;
-    align-items: center;
-
-    .ResultIcon {
-        margin-right: 5px;
-    }
-`
-
-export const ChartContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 50vh;
-    overflow: hidden;
-    canvas {
-        width: 100% !important;
-        height: auto !important;
-    }
-    @media all and (max-width: 800px) {
-        padding: 0;
-    }
-`;
-
-export const ChartControls = styled.div`
-    display: flex;
-    justify-content: flex-end;
-    margin-bottom: 1rem;
-    gap: 1rem;
-    padding: 0 1rem;
-`;
-
-export const ChartSelect = styled.select`
-    padding: 0.5rem 1rem;
-    font-size: 1rem;
-    border: 1px solid ${({ theme }) => theme.borderColor};
-    border-radius: 4px;
-    background-color: ${({ theme }) => theme.backgroundColor};
-    color: ${({ theme }) => theme.textColor};
-    cursor: pointer;
-
-    &:focus {
-        outline: none;
-        box-shadow: 0 0 0 1px #777;
-    }
-
-    @media (max-width: 799px) {
-        font-size: 0.7rem;
-        padding: 0.2rem 0.4rem;
-    }
 `;
